Move page components into "pages" directory.
