+++ frontend/admin/index.html
@@ -14,6 +14,6 @@
 </head>
 <body>
 <div id="root"></div>
-<script src="../src/admin/admin.tsx" type="module"></script>
+<script src="../src/admin/index.tsx" type="module"></script>
 </body>
 </html>
