+++ frontend/src/admin/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
-import AdminApp from './app/AdminApp'
+import AdminApp from './AdminApp'
 import './admin.css'
 
 const container = document.getElementById('root')!
